Migrate filter tags to builders
This is partly so we don't replicate the filter logic and a step towards
normalising all the build steps to use intermediate tools.
